refactor: remove indicator design evaluation method from base indicator class and create a static method in feature evaluator similar to optimize threshold - indicator should just own indicator logic
